update file structure
